#738 confirm before deleting items

+++ codalab/apps/web/static/js/worksheet/table_bundle_interface.js
@@ -161,17 +161,22 @@ var TableBundle = React.createClass({
                 interpreted_row_indexes.push(reactRows[k].props.key);
             }
         }
-        //delete and get our new interpreted. raw is handeled by ws_obj
-        new_interpreted_rows = ws_obj.deleteTableRow(this.state, interpreted_row_indexes);
-        //uncheck so we don't get any weird checked state hanging around
-        this.unCheckRows();
-        // go through and uncheck all the rows to get rid of lingering states
-        this.setState({
-            interpreted: new_interpreted_rows,
-            rowFocusIndex: Math.max(this.state.rowFocusIndex - 1, 0)
-        });
-        // TODO: REMOVE _OWNER
-        this._owner.saveAndUpdateWorksheet();
+        var confirm_string = interpreted_row_indexes.length === 1 ? 'this row?' : interpreted_row_indexes.length + ' rows?'
+        if(interpreted_row_indexes.length && window.confirm("Do you really want to delete " + confirm_string)){
+            //delete and get our new interpreted. raw is handeled by ws_obj
+            new_interpreted_rows = ws_obj.deleteTableRow(this.state, interpreted_row_indexes);
+            //uncheck so we don't get any weird checked state hanging around
+            this.unCheckRows();
+            // go through and uncheck all the rows to get rid of lingering states
+            this.setState({
+                interpreted: new_interpreted_rows,
+                rowFocusIndex: Math.max(this.state.rowFocusIndex - 1, 0)
+            });
+            // TODO: REMOVE _OWNER
+            this._owner.saveAndUpdateWorksheet();
+        } else {
+            return false;
+        }
     },
     saveEditedItem: function(index, interpreted){
         this.props.handleSave(index, interpreted);
